Final tweaks, moved the dropdown jsx and css to its own component folder, fixed paths in imports
