+++ client/src/pages/Profile/Profile.scss
@@ -425,8 +425,8 @@
   justify-content: center;
   align-items: center;
   text-align: center;
-  margin-left: 30px;
-  margin-right: 15px;
+  margin-left: 35px;
+  margin-right: 20px;
   @include devices(tablet) {
     flex-direction: column;
     margin: auto;
